+++ apps/web/pages/_document.tsx
@@ -18,7 +18,6 @@ export default class MyDocument extends Document {
         </Head>
         <body>
           <Main />
-          <div id="modal-root" />
           <NextScript />
         </body>
       </Html>
