+++ apps/web/src/components/common/drawer/search/SearchDrawer.tsx
@@ -8,9 +8,9 @@ import { HASH_TAG_DUMMY_LIST } from '@/assets/dummy/community'
 import { ArrowIcon } from '@/assets/icons'
 import { SearchInput, TagBtn } from '@/components/common'
 
-import { AnimatePortal } from '@/components/common/portal'
+import { MotionPortal } from '@/components/common/portal'
 import { defaultFadeInVariants } from '@/constants/motions'
-interface Props extends ComponentProps<typeof AnimatePortal> {
+interface Props extends ComponentProps<typeof MotionPortal> {
   onClose: () => void
 }
 
@@ -31,7 +31,7 @@ function SearchDrawer({ isMounted, mode, onClose }: Props) {
   const { handleKeyListener } = useArrowKeyTrap(recentKeywordRef, 'row')
 
   return (
-    <AnimatePortal isMounted={isMounted} mode={mode}>
+    <MotionPortal isMounted={isMounted} mode={mode}>
       <m.div css={S.overlayCss} initial="initial" animate="animate" exit="exit" variants={defaultFadeInVariants}>
         <m.div css={S.contentCss} variants={S.searchDrawerVariants}>
           <S.ModalHeader>
@@ -81,7 +81,7 @@ function SearchDrawer({ isMounted, mode, onClose }: Props) {
           </S.Hashtag>
         </m.div>
       </m.div>
-    </AnimatePortal>
+    </MotionPortal>
   )
 }
