+++ apps/web/src/components/common/portal/MotionPortal.tsx
@@ -1,0 +1,18 @@
+import { AnimatePresence } from 'framer-motion'
+import { type ComponentProps } from 'react'
+import Portal from './Portal'
+
+interface Props extends ComponentProps<typeof Portal> {
+  isMounted: boolean
+  mode?: ComponentProps<typeof AnimatePresence>['mode']
+}
+
+function MotionPortal({ children, isMounted, mode = 'wait' }: Props) {
+  return (
+    <Portal>
+      <AnimatePresence mode={mode}>{isMounted && children}</AnimatePresence>
+    </Portal>
+  )
+}
+
+export default MotionPortal
